+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Mailgun for Craft CMS
 
+## Unreleased
+
+### Added
+- Domain, API Key, and Endpoint settings can now be set to environment variables. ([#5](https://github.com/craftcms/mailgun/issues/5)) 
+
 ## 1.3.0 - 2018-11-08
 
 ### Added